Adds conversion comment refresh requests via Private Messages
This is my first attempt at implementing refresh requests via Private
Messages sent to the bot. I am not very familiar with nodejs so I hit a
stumbling block with writing tests, — any guidance on that would be
greatly appreciated. Specifically, the function
`privateMessageFunction` is being called at the start of each Private
Message test but I can’t find what that function does.

- [ ] Add tests
- [ ] Add analytics

Some notes on my implementation:

- Requires 3 API requests to complete the refresh process
- `postTitle` is a required value for the conversion process, however
the comment endpoint does not return the value so I have used `This is
a dummy value.` to allow it to function instead of refactoring the
comment conversion process as to keep my changes small
- Does not reply to the user to confirm the refresh request
- If the ID provided does not return a value from the reddit API then
the refresh request is skipped
- If the comment no longer has anything to convert then no edit is made

If a comment is edited and no longer contains any values that need
converting the comment the bot made could be deleted, or maybe instead
the comment could be edited to have a sassy personality reply,
something referencing the fact that the commenter caused the bot to do
some work and now that work was in vain? I didn’t implement this
because I think it would be a good feature to allow a first timer to
implement because they’d get to have a big impact on the way the bot
acts.